Code cleanup/refactoring. Tab completion implemented.

+++ bin/index.js
@@ -10,16 +10,27 @@ const readline = require('readline');
 const { check, updateLocale } = require("yargs");
 const { exit } = require("process");
 const clear = require('clear');
+// import process from 'node:process';
 
 var bookmarks = new Array();
 var current = "";
-var result = {};
+var matches = {};
+
+const GUI_MODE_RUN = 1;
+const GUI_MODE_DELETE = 2;
+var guiMode = 1;
+
+const LIST_MINIMUM_NAME_WIDTH = 25;
+const LIST_MAXIMUM_COMMAND_WIDTH = 80;
+const LABEL_TITLE = "BLAST - Bookmark management interface";
+const LABEL_TITLE_WIDTH = LABEL_TITLE.length;
+const LIST_NUMBER_COLUMN_WIDTH = 5;
 
 /* ASCII codes for entering or leaving full screen mode */
 const enterAltScreenCommand = '\x1b[?1049h';
 const leaveAltScreenCommand = '\x1b[?1049l';
 
-/* Path to file containing bookmarks */
+/* Path to file containing bookmarks etc. */
 const bookmarksFile = os.homedir() + "/.config/blast/bookmarks.json";
 const commandFile = os.homedir() + "/.config/blast/command.sh";
 
@@ -44,6 +55,8 @@ if (!fs.existsSync(bookmarksFile)) {
 	return;
 }
 
+options.name = options.name.trim();
+
 if (options.action == "save") {
 	/* 
 	* Bookmark command specified at cli
@@ -56,13 +69,16 @@ if (options.action == "save") {
 	}
 
 	let saveBookmark = function() {
-		console.log("Bookmarking last command as " + chalk.yellow.bold(options.name) + ":");
+		console.log("Bookmarking previous history entry as " + chalk.yellow.bold(options.name) + ":");
 		console.log(chalk.bold.green(options.history));
 
 		/* Store bookmark */
 		bookmarks[options.name] = options.history;
+		const ordered = Object.fromEntries(Object.entries(bookmarks).sort(
+			(a, b) => a[0].localeCompare(b[0])
+		));
 
-		let data = JSON.stringify(bookmarks, null, 4);
+		let data = JSON.stringify(ordered, null, 4);
 		fs.writeFileSync(bookmarksFile, data);
 	}
 
@@ -104,17 +120,23 @@ if (options.action == "save") {
 	let found = 0;
 
 	for (let key of Object.keys(bookmarks)) {
-		if ( key.toLowerCase().startsWith(options.name.toLowerCase()) ) {
+		if ( key.toLowerCase().includes(options.name.toLowerCase()) ) {
 			found++;
 
-			console.log("\n    Bookmark " + chalk.yellow.bold(key + ":"));
-			console.log("    " + chalk.bold.green(bookmarks[key]) + "\n");
+			if (found == 1) {
+				process.stdout.write("\n");
+			}
+
+			process.stdout.write("    Bookmark " + chalk.yellow.bold(key + ": "));
+			console.log(chalk.italic(bookmarks[key]));
 		}
 	}
 
 	if (found == 0) {
-		console.log("\n    " + chalk.red("No bookmarks found matching: ") + chalk.yellow.bold(options.name) + "\n");
+		console.log("\n    " + chalk.red("No bookmarks found matching: ") + chalk.yellow.bold(options.name));
 	}
+
+	process.stdout.write("\n");
 } else if (options.action == "use") {
 	/*
 	* Run command specified at cli
@@ -126,12 +148,12 @@ if (options.action == "save") {
 
 		/* Have user confirm */
 		let rl = readline.createInterface(process.stdin, process.stdout);
-		rl.question("Run this command? [yes]/no: ", function (answer) {
+		rl.question("Execute this command? [yes]/no: ", function (answer) {
 			if (answer == "no" || answer == "n") {
 				console.log("Aborted");
 				process.exit(1);
 			} else {
-				console.log("Running command...\n");
+				console.log("Executing command...\n");
 
 				fs.writeFileSync(commandFile, bookmarks[options.name]);
 				rl.close();
@@ -141,8 +163,7 @@ if (options.action == "save") {
 	} else {
 		console.log("\n    " + chalk.red("No bookmark found matching: ") + chalk.yellow.bold(options.name) + "\n");
 	}
-}
-else if (options.action = "showall") {
+} else if (options.action = "showall") {
 	/*
 	 * Show GUI
 	 */
@@ -154,8 +175,11 @@ else if (options.action = "showall") {
 		fs.writeFileSync(commandFile, "");
 		console.log("Bookmarks file does not exist"); // TODO: create file if not found
 	} else {
-		/* Start GUI */
+		/* Draw GUI initially */
+		findMatches();
+		redrawGUI();
 		handleKeyboardEvents();
+		process.stdin.resume();
 	}
 } else {
 	console.log("Error: " + options.status); // TODO: better error handling
@@ -166,9 +190,6 @@ else if (options.action = "showall") {
  * Handle user GUI interaction
  */
 function handleKeyboardEvents() {
-	/* Draw GUI initially */
-	redrawGUI();
-
 	process.stdin.currentLine = '';
 	process.stdin.setRawMode(true);
 
@@ -176,6 +197,12 @@ function handleKeyboardEvents() {
 		const charAsAscii = buf.toString().charCodeAt(0);
 
 		// console.log(charAsAscii);
+
+		/* Avoid exiting when arrow keys pressed */
+		if (buf.toString().length == 3) {
+			let x = buf.toString().charCodeAt(2);
+			if ([65, 66, 67, 68].includes(x)) return;
+		}
 
 		switch (charAsAscii) {
 			case 38:
@@ -190,42 +217,79 @@ function handleKeyboardEvents() {
 				process.kill(process.pid, 'SIGINT');
 				break;
 
+			/* Ctrl-d */
+			case 0x04:
+				if (guiMode == GUI_MODE_DELETE) {
+					guiMode = GUI_MODE_RUN;
+					current = "";
+					redrawGUI();
+				} else {
+					guiMode = GUI_MODE_DELETE;
+					current = "";
+					redrawGUI();
+				}
+
+				break;
+
 			/* Esc */
 			case 27:
+				if (guiMode == GUI_MODE_DELETE) {
+					current = "";
+					guiMode = GUI_MODE_RUN;
+					redrawGUI();
+					break;
+				}
+
 				console.log('Exiting..');
 
 				process.stdout.write(leaveAltScreenCommand);
 				process.kill(process.pid, 'SIGINT');
+
 				break;
 			
 			/* Backspace */
 			case 127:
 				current = current.slice(0, -1);
+
+				findMatches();
 				redrawGUI();
 
 				break;
 
 			/* Tab */
 			case 0x09:
+				autoComplete(); // fs.appendFileSync("/home/x/debug.log", "tab current = " + current + "\n");
+				findMatches();
 				redrawGUI();
+
 				break;
 
 			/* Enter */
 			case 13:
-				let keys = Object.keys(result);
-				if (keys.length == 1) {
-					console.log("\nRunning command " + chalk.yellow.bold(keys[0]) + ": " + chalk.green(bookmarks[keys[0]]) + "\n");
+				let exactMatchIndex = Object.values(matches).indexOf(2);
+				if (exactMatchIndex > -1) {
+					let commandName = Object.keys(matches)[exactMatchIndex];
 
 					process.stdout.write(leaveAltScreenCommand);
-					fs.writeFileSync(commandFile, bookmarks[keys[0]]);
+					fs.writeFileSync(commandFile, bookmarks[commandName]);
+					console.log("\nExecuting command " + chalk.yellow.bold(commandName) + ": " + chalk.italic(bookmarks[commandName]) + "\n");
+
 					exit();
 				}
 
 				break;
-
+						
 			/* User input */
 			default:
-				current += String.fromCharCode(charAsAscii);
+				if (guiMode == GUI_MODE_DELETE) {
+					if (is_numeric(String.fromCharCode(charAsAscii))) {
+						current += String.fromCharCode(charAsAscii);
+					}
+				} else {
+					current += String.fromCharCode(charAsAscii);
+				}
+
+				findMatches();
 				redrawGUI();
 
 				break;
@@ -233,85 +297,182 @@ function handleKeyboardEvents() {
 	});
 }
 
+/**
+ * Keeps skipping forward in command names until 
+ */
+function autoComplete() {
+	let nextCharacters = {};
+
+	let max = current.length;
+	for (let key of Object.keys(bookmarks)) {
+		if ( key.toLowerCase().startsWith(current.toLowerCase()) ) {
+			max = Math.max(key.length, max);
+		}
+	}
+
+	for (let key of Object.keys(bookmarks)) {
+		if ( key.toLowerCase().startsWith(current.toLowerCase()) ) {
+			key = key.padEnd(max, " ");
+
+			if (key.length > current.length) {
+				let keyAfter = key.substring(current.length, current.length + 1);
+				nextCharacters[keyAfter] = keyAfter;
+			}
+		}
+	}
+
+	if (Object.keys(nextCharacters).length == 1) {
+		current += nextCharacters[Object.keys(nextCharacters)[0]]
+		autoComplete();
+	} else {
+		return;
+	}
+}
+
+/**
+ * Update datastructure with match states
+ */
+function findMatches() {
+	for (let key of Object.keys(bookmarks)) {
+		if ( current != "" && key.toLowerCase().startsWith(current.toLowerCase()) ) {
+			key.toLowerCase() == current.toLowerCase() ? matches[key] = 2 : matches[key] = 1;
+		} else {
+			matches[key] = 0;
+		}
+	}
+}
 
 /**
  * Redraw GUI after user interaction
  */
 function redrawGUI() {
-	result = {};
+	exactMatch = "";
 
-	const match = chalk.green.bold;
+	const match = chalk.white.italic;
+	const matchUnique = chalk.green;
 	const noMatch = chalk.white;
 	const tableBorder = chalk.white;
 	const usage = chalk.white;
 
+	/* Prepare/clear screen */
 	readline.cursorTo(process.stdout, 0, 0);
 	readline.clearScreenDown(process.stdout); // process.stdout.write("\x1Bc")
 	readline.cursorTo(process.stdout, 0, 60);
 
-	let maxKeyLength = 0;
+	/* Find width of name column */
+	let nameColumnWidth = 0;
 	for (let key of Object.keys(bookmarks)) {
-		maxKeyLength = Math.max(key.length, maxKeyLength);
+		nameColumnWidth = Math.max(key.length, nameColumnWidth);
 	}
+	nameColumnWidth = Math.max(nameColumnWidth, LIST_MINIMUM_NAME_WIDTH); // Minimum 25 characters wide
 
-	maxKeyLength = Math.max(maxKeyLength, 25);
+	/* Find available space for command and width of table */
+	let remainingWidth = terminalWidth() - LIST_NUMBER_COLUMN_WIDTH - nameColumnWidth - 7; // 7 = padding and borders
+	let commandWidth = Math.max( Math.min(remainingWidth, LIST_MAXIMUM_COMMAND_WIDTH), 0 );
 
-	console.log( chalk.yellow.bold("Usage\n") );
+	/* Title and menu */
+	let menuWidth = (" ^D Delete " + "^C/Esc Exit").length;
+	let menu = chalk.bgWhite.black(" ^D ") + " Delete " + chalk.bgWhite.black(" ^C/Esc ") + " Exit";
+	let tableWidth = nameColumnWidth + commandWidth + LIST_NUMBER_COLUMN_WIDTH + 4;
+
+	process.stdout.write( chalk.yellow.bold(LABEL_TITLE) );
+	if (menuWidth + LABEL_TITLE_WIDTH + 4 > terminalWidth()) process.stdout.write( "\n" );
+	readline.cursorTo(process.stdout, tableWidth - menuWidth);
+	process.stdout.write( menu );
+	process.stdout.write( "\n" );
+
+	/* Usage info */
+	console.log( "\n" + chalk.yellow("Usage:") );
 	console.log( usage("blast as ") + chalk.blue("'bookmark-name'") + "     <- Save bookmark");
-	console.log( usage("blast ") + chalk.blue("'bookmark-name'") + "        <- Run bookmarked command");
+	console.log( usage("blast ") + chalk.blue("'bookmark-name'") + "        <- Execute bookmarked command");
 	console.log( usage("blast delete ") + chalk.blue("'bookmark-name'") + " <- Delete bookmark" );
-	console.log( usage("blast show ") + chalk.blue("'bookmark-name'") + "   <- List bookmarks starting with specified string");
+	console.log( usage("blast show ") + chalk.blue("'bookmark-name'") + "   <- List bookmarks starting with specified string\n");
 
-	process.stdout.write( "\n" + chalk.bgWhite.black(" ^C / Esc ") + " Exit" + "\n" );
+	/* Draw top border of table */
+	process.stdout.write( tableBorder("╭────┬" + "─".repeat(nameColumnWidth + 2)) );
+	process.stdout.write( tableBorder("┬" + "─".repeat(commandWidth + 2) + "╮\n") );
 
-	process.stdout.write( tableBorder("╭") );
-	process.stdout.write( tableBorder("─".repeat(maxKeyLength + 2)) );
-	process.stdout.write( tableBorder("┬") );
-	process.stdout.write( tableBorder("─".repeat(80)) );
-	process.stdout.write( tableBorder("╮\n") );
+	/* Draw line for each bookmark in table */
+	for (let key of Object.keys(matches)) {
+		let commandName = key.padEnd(nameColumnWidth, " ");
+		let command = bookmarks[key].padEnd(commandWidth, " ");
 
-	for (let key of Object.keys(bookmarks)) {
-		let commandAlias = key.padEnd(maxKeyLength, " ");
-		let command = bookmarks[key].padEnd(78, " ");
+		switch (matches[key]) {
+			case 0:
+				commandName = noMatch(commandName);
+				command = noMatch(command);
+					break;
 
-		if ( current != "" && key.toLowerCase().startsWith(current.toLowerCase()) ) {
-			result[key] = bookmarks[key];
-			commandAlias = match(commandAlias);
-			command = match(command);
-		} else {
-			commandAlias = noMatch(commandAlias);
-			command = noMatch(command);
+			case 1:
+				commandName = match(commandName);
+				command = match(command);
+				break;
+
+			case 2:
+				exactMatch = key;
+				commandName = matchUnique(commandName);
+				command = matchUnique(command);
+				break;
 		}
 
-		process.stdout.write( tableBorder("│ ") );
-		process.stdout.write(commandAlias);
-		process.stdout.write( tableBorder(" │ ") );
-		process.stdout.write(command);
-		process.stdout.write(" │\n");
+		let numberColumn = (Object.keys(bookmarks).indexOf(key) + 1).toString().padStart(2, " ");
 
-		process.stdout.write( tableBorder("│") );
-		process.stdout.write( tableBorder("╌".repeat(maxKeyLength + 2)) );
-		process.stdout.write( tableBorder("┼") );
-		process.stdout.write( tableBorder("╌".repeat(80)) );
-		process.stdout.write( tableBorder("│\n") );
-	
+		/* Bookmark line */
+		process.stdout.write( tableBorder("│ ") + numberColumn + tableBorder(" │ ") );
+		process.stdout.write( commandName + tableBorder(" │ ") + command + tableBorder(" │\n") );
+
+		/* Divider, but omit if last bookmark in list */
+		if (key != Object.keys(bookmarks).at(-1)) {
+			process.stdout.write( tableBorder("│╌╌╌╌│" + "╌".repeat(nameColumnWidth + 2)) );
+			process.stdout.write( tableBorder("┼" + "╌".repeat(commandWidth + 2) + "│\n") );
+		}	
 	}
 
-	process.stdout.write( tableBorder("╰") );
-	process.stdout.write( tableBorder("─".repeat(maxKeyLength + 2)) );
-	process.stdout.write( tableBorder("┴") );
-	process.stdout.write( tableBorder("─".repeat(80)) );
-	process.stdout.write( tableBorder("╯\n") );
+	/* Draw bottom border of table */
+	process.stdout.write( tableBorder("╰────┴" + "─".repeat(nameColumnWidth + 2)) );
+	process.stdout.write( tableBorder("┴" + "─".repeat(commandWidth + 2) + "╯\n") );
 
 	/* Update input line */
-	if (Object.keys(result).length == 1) {
-		process.stdout.write(chalk.white(": "));
-		process.stdout.write(chalk.yellow.bold(current));
+	if (exactMatch != "") {
+		let command = bookmarks[exactMatch]
+		// console.log("\n  " + chalk.green.bold(command));
+		console.log(chalk.green("\nMatch found: press enter to execute command"));
+		process.stdout.write(": " + current);
 	} else {
-		process.stdout.write(chalk.reset(": " + current));
+		if (guiMode == GUI_MODE_RUN) {
+			console.log(chalk.yellow("\nSpecify name of command to execute:"));
+			// process.stdout.write(chalk.reset(": " + current));
+			process.stdout.write(": " + current);
+		} else {
+			console.log(chalk.yellow("\nSpecify number of command to delete and press enter:"));
+			process.stdout.write(": " + current);
+		}
 	}
 }
 
-function checkIfStringStartsWith(str, substrs) {
-	return substrs.some(substr => str.toLowerCase().startsWith(substr.toLowerCase()));
+process.on('SIGWINCH', () => {
+	redrawGUI();
+});
+
+function is_numeric(string) {
+	return /^\d+$/.test(string);
 }
+
+function terminalWidth() {
+	return process.stdout.getWindowSize()[0];
+}
+
+
+// /* Arrow keys */
+// case 28:
+// 	break;
+// case 0x4B:
+// 	break;
+// case 0x4D:
+// 	break;
+// case 0x48:
+// 	break;
+
+// Object.keys(matches).forEach(function(value, index, array) {
+// 	fs.appendFileSync("/home/x/debug.log", value + " = " + matches[value] + "\n");
+// });
